Split up block implementation file
